Add support for newtype structs
This change adds support for newtype structs to the crate. There
arguably is no "variant" name in a stricter sense to retrieve there, but
at the same time there should be no harm in returning the struct name --
which really is the only sensible thing a user could ask for.